moved classes because I was doing the wrong one
