+++ .gitignore
@@ -2,3 +2,4 @@
 /shelf/
 /workspace.xml
 /Main.java
+/PrefixCalculator.java
