+++ PrefixCalculator.java
@@ -1,2 +1,6 @@
 public class PrefixCalculator extends Calculator{
+    @Override
+    public void calculate(){
+
+    }
 }
